Piggybacking on this pull request as during reviewing this request there were other issues in how the templating, and remote resourcing was being handled that this commit fixes.
Updated template directory for the pyvis package to import the template directory as a jinja template environment instead of by file. Previous was imported by filename and didn't allow for include statements as well as not being a template directory.
Sources were added without integrety hashes. Updated to import using said integrity checks.
Added new argument to Network declration to change how resources are packaged into the end product. This adds remote, in_line, and local as options for building the template. In-line takes raw js files from the lib directory and injects them as inline javascript and can be used for local non-internet access situations, as well when remote files don't wish to be trusted. Remote pulls said sources from remote with integrety checks handled. local will pull the resources from the local lib directory as was being performed before.
in_line and remote fixes an issue with pyvis in jupyter-notebook on chrome and safari that caused graphics to error with a blank object due to jupyter denying access due to a CORBS policy due to chrome and safari removing referer on requests to same origin cdns.
<meta charset='utf-8'> is a fix for the tom-select pacakge as there is non utf-8 characters implimented.
Wrapped tom_select template initialization to check if select_menu is ture as they were being initialized and thus causing errors when this option was set to false.
Updated jupyter-notebook to contain options, and examples of the node selection to be able to test changes.